requirements.txt
streamlit==1.56.0
openai==2.30.0
Pillow==12.2.0